Nest Templater class within Staging
This class is now pretty trivial and is Staging-specific, so just having
it live here is simpler and clearer.
