+++ src/staging.rb
@@ -1,9 +1,11 @@
 # frozen_string_literal: true
 
+require 'active_support/core_ext/string/strip'
+require 'active_support/core_ext/string/filters'
+
 require 'data'
 require 'file_path'
 require 'recursive-open-struct'
-require 'templater'
 require 'managed_file'
 
 module Metalware
@@ -102,6 +104,27 @@ module Metalware
         managed_file_content_args.push(comment_char: data.comment_char)
       end
       ManagedFile.content(*managed_file_content_args)
+    end
+
+    class Templater
+      def initialize(staging)
+        @staging = staging
+      end
+
+      def render(
+        namespace,
+        template,
+        sync_location,
+        dynamic: {},
+        **staging_options
+      )
+        rendered = namespace.render_file(template, **dynamic)
+        staging.push_file(sync_location, rendered, **staging_options)
+      end
+
+      private
+
+      attr_reader :staging
     end
   end
 end
